+++ Src/AutoFixture/IRequestMemberTypeResolver.cs
@@ -10,9 +10,6 @@ namespace AutoFixture
         /// <summary>
         /// Tries to determine requested member type and returns it.
         /// </summary>
-        /// <param name="request">Relay request object</param>
-        /// <param name="memberType">Output variable for resolved request member type</param>
-        /// <returns></returns>
         bool TryGetMemberType(object request, out Type memberType);
     }
 }
